+++ components/src/main/java/com/mercadolibre/android/andesui/snackbar/factory/AndesSnackbarConfigurationFactory.kt
@@ -1,6 +1,7 @@
 package com.mercadolibre.android.andesui.snackbar.factory
 
 import android.content.Context
+import android.util.Log
 import android.view.View
 import com.mercadolibre.android.andesui.R
 import com.mercadolibre.android.andesui.color.AndesColor
@@ -16,7 +17,8 @@ internal data class AndesSnackbarAttrs(
     var andesSnackbarType: AndesSnackbarType,
     var andesSnackbarText: String,
     var andesSnackbarDuration: AndesSnackbarDuration,
-    var andesSnackbarAction: AndesSnackbarAction? = null
+    var andesSnackbarAction: AndesSnackbarAction? = null,
+    var andesSnackbarErrorCode: String? = null
 )
 
 internal data class AndesSnackbarConfiguration(
@@ -33,6 +35,7 @@ internal data class AndesSnackbarConfiguration(
 internal object AndesSnackbarConfigurationFactory {
 
     fun create(context: Context, view: View, andesSnackbarAttrs: AndesSnackbarAttrs): AndesSnackbarConfiguration {
+        resolveType(andesSnackbarAttrs)
         return AndesSnackbarConfiguration(
                 view = view,
                 backgroundColor = resolveBackgroundColor(andesSnackbarAttrs.andesSnackbarType.type),
@@ -43,6 +46,13 @@ internal object AndesSnackbarConfigurationFactory {
                 marginRight = context.resources.getDimension(R.dimen.andes_snackbar_right_margin).toInt(),
                 marginBottom = context.resources.getDimension(R.dimen.andes_snackbar_bottom_margin).toInt()
         )
+    }
+
+    private fun resolveType(andesSnackbarAttrs: AndesSnackbarAttrs){
+        if (!andesSnackbarAttrs.andesSnackbarErrorCode.isNullOrEmpty() && andesSnackbarAttrs.andesSnackbarType != AndesSnackbarType.ERROR) {
+            andesSnackbarAttrs.andesSnackbarType = AndesSnackbarType.ERROR
+            Log.e("Andes UI", "Type not supported for errorCode")
+        }
     }
 
     private fun resolveBackgroundColor(type: AndesSnackbarTypeInterface) = type.primaryColor()